requirements.txt
alembic==0.9.9
amqp==2.2.2
aniso8601==3.0.0
attrs==18.1.0
autopep8==1.3.5
billiard==3.5.0.3
bleach==2.1.3
click==6.7
coverage==4.5.1
croniter==0.3.22
flake8==3.5.0
Flask==1.0.2
flask-buzz==0.1.5
Flask-Cors==3.0.4
Flask-Migrate==2.1.1
flask-praetorian==0.4.4
flask-restplus==0.10.1
Flask-RQ2==18.0
Flask-SQLAlchemy==2.3.2
html5lib==1.0.1
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
kombu==4.2.0
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.1
mccabe==0.6.1
more-itertools==4.1.0
parso==0.2.0
passlib==1.7.1
pendulum==2.0.1
pluggy==0.6.0
py==1.5.3
py-buzz==0.3.4
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.1
pytest==3.5.1
python-dateutil==2.7.2
python-dotenv==0.8.2
python-editor==1.0.3
python-http-client==3.0.0
pytz==2018.4
pytzdata==2018.5
redis==2.10.6
rq==0.10.0
rq-scheduler==0.8.3
sendgrid==5.3.0
six==1.11.0
SQLAlchemy==1.2.7
vine==1.1.4
webencodings==0.5.1
Werkzeug==0.14.1
yapf==0.21.0
